% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VUBtemplates.R
\name{article}
\alias{article}
\title{Article}
\usage{
article(..., md_extensions = c("-autolink_bare_uris"))
}
\arguments{
\item{...}{Optional arguments passed on to \code{\link{pdf_document}}}

\item{md_extensions}{Markdown Extensions, cf. \code{\link{pdf_document}}}
}
\description{
Formatting an article in th VUB style
}
\details{
The function is called by the Markdown model Article
}
