% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VUBtemplates.R
\name{Knit}
\alias{Knit}
\alias{KnitAll}
\alias{KnitArticle}
\title{Knit}
\usage{
KnitAll(destination = "docs")

KnitArticle(destination = "docs")
}
\arguments{
\item{destination}{Destionation file of the documents}
}
\description{
Creation of all the documents from the models
}
\details{
Used to verify that the models function well
}
